ark-core/runtime/requirements.txt
nats-py==2.9.0
requests==2.32.3
duckdb==1.1.3
